Renamed 'tasks' to 'assignedtasks' and fixed endpoint along with using taskUtil
